Continued work to get WebExt/Redux going
