+++ src/constants.js
@@ -1,2 +1,3 @@
 export const ADD_CLASSIFICATION = 'ADD_CLASSIFICATION';
 export const ANALYZE_CONTENT = 'ANALYZE_CONTENT';
+export const SET_URL = 'SET_URL';
