+++ src/popup.js
@@ -1,132 +1,23 @@
-import 'material-icons-font/material-icons-font.css';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import MainView from './MainView';
 
-import {connect} from 'react-redux'
 import {Provider} from 'react-redux';
 import {createUIStore} from 'redux-webext';
 
-class Popup extends React.Component {
-  constructor(props) {
-    super(props);
-    console.log('Bubble popup');
-
-    this.state = {
-        tag: '',
-        matchedTag: '',
-        activeTab: null
-    };
-
-    this.handleChange = this.handleChange.bind(this);
-    this.classify = this.classify.bind(this);
-    var report = this.report = this.report.bind(this);
-
-    var port = browser.runtime.connect({name:"bubble-scan"});
-    port.onMessage.addListener(report);
-
-    browser.tabs.executeScript({
-          code: 'infobubble.analyze()'
-    });
-  }
-
-  componentDidMount() {
-    // Get the active tab and store it in component state.
-    chrome.tabs.query({active: true}, tabs => {
-      this.setState({activeTab: tabs[0]});
-    });
-  }
-
-  report(m) {
-    console.log('Report!', m);
-    if (m.action === 'REPORT') {
-        this.setState({ matchedTag: m.tag });
-    }
-  }
-
-  analyzeText() {
-      console.log('Analyze');
-      browser.tabs.executeScript({
-          code: 'infobubble.analyze()'
-      }).then(function(result) {
-          console.log('Success!', result);
-          result.forEach(function(item) {
-              console.log('Result item:', item);
-          });
-      }, function(error) {
-          console.log('Error!', error);
-      });
-  }
-
-  classify(event) {
-      var tag = this.state.tag;
-
-      browser.tabs.executeScript({
-          code: 'infobubble.classify("' + tag + '")'
-      }).then(function(result) {
-          console.log('Success!', result);
-          result.forEach(function(item) {
-              console.log('Result item:', item);
-          });
-      }, function(error) {
-          console.log('Error!', error);
-      });
-
-      event.preventDefault();
-  }
-
-  handleChange(ev) {
-    this.setState({tag: ev.target.value});
-  }
-
-  render() {
-    console.log(this);    
-    const {url} = this.props;
-    return (
-            <div>
-            <MuiThemeProvider>
-                <MainView classification={this.state.matchedTag} url={url ? url : 'loading...'} />
-            </MuiThemeProvider>
-            <hr/>
-
-            <button onClick={this.analyzeText}>Analyze</button>
-
-            <form onSubmit={this.classify}>
-                <input type="text" value={this.state.tag} onChange={this.handleChange} />
-                <button onClick={this.classify}>Classify</button>
-            </form>
-        </div>
-    );
-  }
-}
-
-const mapStateToProps = (state) => {
-    console.log('Map state to props ', state);
-    return {
-        url: state.url 
-    }
-}
-
-const mapDispatchToProps = (dispatch) => {
-    return {
-    }
-}
-
-const MyPopup = connect(mapStateToProps, mapDispatchToProps)(Popup)
-
 async function initApp() {
     const store = await createUIStore();
-
     console.log('store', store);
 
     const mountNode = document.getElementById('app');
-//    document.body.appendChild(mountNode);
 
     ReactDOM.render(
-        <Provider store={store}>
-            <MyPopup/>
-        </Provider>,
+        <MuiThemeProvider>
+            <Provider store={store}>
+                <MainView/>
+            </Provider>
+        </MuiThemeProvider>,
         mountNode
     );
 }
