+++ src/store.js
@@ -1,15 +1,16 @@
 import {createStore} from 'redux';
 import {createBackgroundStore} from 'redux-webext';
-import {ADD_CLASSIFICATION, ANALYZE_CONTENT} from './constants';
+import {ADD_CLASSIFICATION, SET_URL, ANALYZE_CONTENT} from './constants';
 import reducer from './reducers';
-import {addClassification, analyzeContent} from './actions';
+import {addClassification, setUrl, analyzeContent} from './actions';
 
 const store = createStore(reducer);
 
 export default createBackgroundStore({
     store,
     actions: {
-        ADD_CLASSIFICATION: addClassification,
-        ANALYZE_CONTENT: analyzeContent
+        'UI_ADD_CLASSIFICATION': addClassification,
+        'UI_SET_URL': setUrl,
+        'UI_ANALYZE_CONTENT': analyzeContent
     }
 });
